Add files via upload
The STATS+STORIES podcast is celebrating its 300th episode and 10-year anniversary by hosting a data visualization contest. Contestants are invited to create a compelling data visualization that tells a story using the provided dataset, which contains information on over 280 episodes of the podcast. The goal is to explore and uncover insights about the podcast, such as which categories have the most listens, how the program has changed over time, or which episodes stand out as unusual.